Configuration and README updates for animation.
