+++ include/definitions/Defaults.h
@@ -81,3 +81,23 @@
 #ifndef LEDS_BUTTON_14
 #define LEDS_BUTTON_14  -1
 #endif
+
+#ifndef LEDS_BASE_ANIMATION
+#define LEDS_BASE_ANIMATION "RAINBOW"
+#endif
+
+#ifndef LEDS_BASE_ANIMATION_FIRST_PIXEL
+#define LEDS_BASE_ANIMATION_FIRST_PIXEL 0
+#endif
+
+#ifndef LEDS_BASE_ANIMATION_LAST_PIXEL 
+#define LEDS_BASE_ANIMATION_LAST_PIXEL 11
+#endif
+
+#ifndef LEDS_RAINBOW_CYCLE_TIME 
+#define LEDS_RAINBOW_CYCLE_TIME 500
+#endif
+
+#ifndef LEDS_STATIC_COLOR_COLOR
+#define LEDS_STATIC_COLOR_COLOR 255
+#endif
