+++ include/definitions/OpenStickBoard.h
@@ -27,6 +27,9 @@
 #define PIN_BUTTON_13   28
 #define PIN_BUTTON_14   18
 
+#define BOARD_LEDS_PIN 14
+#define BOARD_LEDS_COUNT 16
+
 #define LEDS_DPAD_UP     3
 #define LEDS_DPAD_DOWN   1
 #define LEDS_DPAD_LEFT   0
@@ -39,9 +42,9 @@
 #define LEDS_BUTTON_06   6
 #define LEDS_BUTTON_07   11
 #define LEDS_BUTTON_08   10
-#define DEFAULT_SOCD_MODE SOCD_MODE_NEUTRAL
 
-#define BOARD_LEDS_PIN 14
-#define BOARD_LEDS_COUNT 16
+#define LEDS_BASE_ANIMATION "STATIC"
+
+#define DEFAULT_SOCD_MODE SOCD_MODE_NEUTRAL
 
 #endif
