+++ lib/AnimationStation/src/AnimationStation.hpp
@@ -12,7 +12,7 @@ public:
   AnimationStation(int numPixels);
   void Animate();
   void SetStaticColor(bool defaultAnimation, uint32_t color, int firstPixel, int lastPixel);
-  void SetRainbow(bool defaultAnimation, int firstPixel, int lastPixel);
+  void SetRainbow(bool defaultAnimation, int firstPixel, int lastPixel, int cycleTime);
   void Clear();
 
   uint32_t frame[100];
