+++ lib/Helpers/library.json
@@ -1,0 +1,13 @@
+{
+	"name": "Helpers",
+	"version": "0.0.1",
+	"description": "Nonsense Helpers for RP2040",
+	"keywords": "c c++ baremetal nonsense",
+	"authors": [
+		{
+			"name": "Jonathan Barket",
+			"url": "https://jonathanbarket.com"
+		}
+	],
+	"license": "MIT"
+}
