+++ lib/NeoPico/src/NeoPico.cpp
@@ -17,13 +17,6 @@ void NeoPico::PutPixel(uint32_t pixel_grb) {
     pio_sm_put_blocking(pio0, 0, pixel_grb << 8u);
 }
 
-uint32_t NeoPico::RGB(uint8_t r, uint8_t g, uint8_t b) {
-    return
-            ((uint32_t) (r) << 8) |
-            ((uint32_t) (g) << 16) |
-            (uint32_t) (b);
-}
-
 NeoPico::NeoPico(int ledPin, int numPixels) {
     PIO pio = pio0;
     int sm = 0;
