+++ lib/NeoPico/src/NeoPico.hpp
@@ -11,7 +11,6 @@ public:
   void Clear();
   void SetPixel(int pixel, uint32_t color);
   void SetFrame(uint32_t newFrame[100]);
-  uint32_t RGB(uint8_t r, uint8_t g, uint8_t b);
 private:
   void PutPixel(uint32_t pixel_grb);
   PIO pio = pio0;
